+++ src/Module.php
@@ -7,7 +7,7 @@ namespace Hirtz\Shopify;
 use Override;
 use Yii;
 
-class Module extends \yii\base\Module
+class Module extends \Hirtz\Skeleton\Base\Module
 {
     public ?string $shopifyShopName = null;
     public ?string $shopifyShopDomain = null;
@@ -39,7 +39,7 @@ class Module extends \yii\base\Module
         $this->shopifyShopName ??= Yii::$app->params['shopifyShopName'] ?? null;
 
         $this->shopifyShopDomain ??= Yii::$app->params['shopifyShopDomain'] ?? "$this->shopifyShopName.myshopify.com";
-        $this->shopifyShopDomain = rtrim((string)preg_replace('(^https??//)', '', (string)$this->shopifyShopDomain), '/');
+        $this->shopifyShopDomain = rtrim((string)preg_replace('(^https?://)', '', (string)$this->shopifyShopDomain), '/');
 
         $this->shopifyApiKey ??= Yii::$app->params['shopifyApiKey'] ?? null;
         $this->shopifyApiSecret ??= Yii::$app->params['shopifyApiSecret'] ?? null;